Problem: sets that use an alias might be valid in the DSL but not in the IR
An example is (b #{(inc 42) (successor 42)}) which will yield an exception because there is a duplicate key (the IR).
This is not really a bug. However, design-wise it is nice to see that the expression remains valid.
Other semantic duplicates cannot be caught at this point either.
Relevant ticket: #10

Solution: In the pre-processing step, transform set-literals to (hash-set e1 e2 ...)