Problem: B macro can result in method code too large error
Solution: implement a bb fn that does the rewriting as a function. It takes the code syntax quoted.
Part of the solution extracts the mapping of DSL symbol to function from the B macro and puts it in a map.
pre-process-lisb is streamlined to ignore the namespaces of symbols such as <-- or |.